Extract shared event handling logic to BaseEndScene
This removes duplicated code from the `handle_event` methods in `DefeatScene` and `VictoryScene` by consolidating it into a common `BaseEndScene` class. It also cleans up unused `pygame` imports.
