+++ command_line_conflict/scenes/victory.py
@@ -1,34 +1,10 @@
 import math
 
-import pygame
+from .base import BaseEndScene
 
 
-class VictoryScene:
+class VictoryScene(BaseEndScene):
     """A scene to display when the player wins the game."""
-
-    def __init__(self, game):
-        """Initializes the VictoryScene.
-
-        Args:
-            game: The main game object.
-        """
-        self.game = game
-        self.font = game.font
-        self.time = 0.0
-
-    def handle_event(self, event):
-        """Handles events for the victory scene.
-
-        Args:
-            event: The pygame event to handle.
-        """
-        if event.type == pygame.MOUSEMOTION:
-            pygame.mouse.set_cursor(pygame.SYSTEM_CURSOR_HAND)
-        elif event.type == pygame.KEYDOWN:
-            if event.key in (pygame.K_RETURN, pygame.K_ESCAPE):
-                self.game.scene_manager.switch_to("menu")
-        elif event.type == pygame.MOUSEBUTTONDOWN:
-            self.game.scene_manager.switch_to("menu")
 
     def update(self, dt):
         """Updates the victory scene.
